web/requirements.txt
asn1crypto==0.24.0
attrdict==2.0.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.6.1
cytoolz==0.9.0.1
Django==2.2.1
django-filter==2.1.0
djangorestframework==3.9.4
eth-abi==1.3.0
eth-account==0.3.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.2
eth-rlp==0.1.2
eth-typing==2.1.0
eth-utils==1.6.0
Flask==1.0.3
hexbytes==0.1.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
lru-dict==1.1.6
Markdown==3.1
MarkupSafe==1.1.1
parsimonious==0.8.1
pycparser==2.19
pycryptodome==3.8.1
pytz==2019.1
requests==2.22.0
rlp==1.1.0
six==1.12.0
sqlparse==0.3.0
toolz==0.9.0
urllib3==1.25.2
web3==4.9.2
websockets==6.0
Werkzeug==0.15.4
